# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Video Tools API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/video-projects/{videoProjectId}/video-assets": {
      "get": {
        "summary": "List a project's video assets",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "videoProjectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "videoProjectId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "parentVideoAssetId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "assetType": {
                            "type": "string",
                            "enum": [
                              "source_recording",
                              "edited_master",
                              "short_source_clip",
                              "rendered_short"
                            ]
                          },
                          "uri": {
                            "type": "string"
                          },
                          "durationS": {
                            "type": "number",
                            "nullable": true
                          },
                          "width": {
                            "type": "integer",
                            "nullable": true
                          },
                          "height": {
                            "type": "integer",
                            "nullable": true
                          },
                          "fps": {
                            "type": "number",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "videoProjectId",
                          "parentVideoAssetId",
                          "assetType",
                          "uri",
                          "durationS",
                          "width",
                          "height",
                          "fps",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload a source recording (multipart/form-data, field: file)",
        "tags": [
          "video"
        ],
        "description": "Streams the upload to disk, probes it with ffprobe, and records a source_recording asset.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "videoProjectId",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "videoProjectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "parentVideoAssetId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "assetType": {
                      "type": "string",
                      "enum": [
                        "source_recording",
                        "edited_master",
                        "short_source_clip",
                        "rendered_short"
                      ]
                    },
                    "uri": {
                      "type": "string"
                    },
                    "durationS": {
                      "type": "number",
                      "nullable": true
                    },
                    "width": {
                      "type": "integer",
                      "nullable": true
                    },
                    "height": {
                      "type": "integer",
                      "nullable": true
                    },
                    "fps": {
                      "type": "number",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "videoProjectId",
                    "parentVideoAssetId",
                    "assetType",
                    "uri",
                    "durationS",
                    "width",
                    "height",
                    "fps",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/video-assets/{videoAssetId}": {
      "get": {
        "summary": "Get a video asset",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "videoAssetId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "videoProjectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "parentVideoAssetId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "assetType": {
                      "type": "string",
                      "enum": [
                        "source_recording",
                        "edited_master",
                        "short_source_clip",
                        "rendered_short"
                      ]
                    },
                    "uri": {
                      "type": "string"
                    },
                    "durationS": {
                      "type": "number",
                      "nullable": true
                    },
                    "width": {
                      "type": "integer",
                      "nullable": true
                    },
                    "height": {
                      "type": "integer",
                      "nullable": true
                    },
                    "fps": {
                      "type": "number",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "videoProjectId",
                    "parentVideoAssetId",
                    "assetType",
                    "uri",
                    "durationS",
                    "width",
                    "height",
                    "fps",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a video asset and its file",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "videoAssetId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": [
                    "null"
                  ],
                  "nullable": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/video-assets/{videoAssetId}/file": {
      "get": {
        "summary": "Stream the asset's file (supports HTTP Range)",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "videoAssetId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "206": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "416": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": [
                    "null"
                  ],
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000"
    }
  ],
...
}